[FE-67] refactor: HistoryPage 기능 단위 훅·컴포넌트 분리
- hooks/: useHistoryTarget(앱·배포 선택), useReleaseHistory(이력·지표·페이지), useReleaseToggles(disabled/mandatory 토글)
- components/: ReleaseCard, HistoryPagination 분리 (formatBytes/formatTime은 ReleaseCard로 이동)
- types.ts로 응답 타입 분리
- useEffect 3개 기명화: selectFirstAppOnLoad / selectDefaultDeployment / resetPageOnTargetChange
- 동작 동일, 내부 구조만 변경
